Revert changelog-check to using PR body.
I forgot that we're going to be running changelog-check against
magic-modules, which won't have the changelog file. Instead, the
changelog entries will be in the body, and the magician will generate
the changelog files on the downstream before it gets merged. So rather
than having changelog-check look for the files, let's have it look in
the PR body, which is also much simpler and less code, which is nice.